requirements.txt
asgiref==3.8.1
audioop-lts==0.2.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
defusedxml==0.7.1
dj-rest-auth==7.0.1
django==5.2
django-allauth==0.61.1
django-cors-headers==4.4.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
dotenv==0.9.9
google-auth==2.40.1
google-auth-oauthlib==1.2.2
gunicorn==23.0.0
idna==3.10
markdown==3.7
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pyjwt==2.9.0
pymysql==1.1.1
python-dotenv==1.1.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.7.1
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
typing-extensions==4.13.2
urllib3==2.4.0
wheel==0.40.0
whitenoise==6.8.2
